Change `id` type from `Long` to `Int` in `ProviderConverter`.

+++ composeApp/src/androidMain/kotlin/com/currency/rateman/api/kurzyCz/ProviderConverter.kt
@@ -32,7 +32,7 @@ object ProviderConverter {
 
     fun toProvider(entity: ProviderEntity, rates: List<CurrencyRateEntity>): Provider {
         return Provider(
-            id = entity.id.toLong(),
+            id = entity.id,
             name = entity.name,
             baseCurrency = CurrencyCode.valueOf(entity.baseCurrency),
             rates = rates.map { rate ->
